refactor: consistent code location and naming for DbPut
